Move tooltip stories to subfolder
